#11 [feat] 키워드 일치 개수 조회, 매칭되면 안되는 키워드 존재 여부, 구역 추천 메소드를 공통메소드로 분리한다.
