+++ src/main/java/kusitms/backend/stadium/domain/entity/enums/JamsilStadiumStatusType.java
@@ -10,7 +10,7 @@ import java.util.*;
 @Slf4j
 @Getter
 @RequiredArgsConstructor
-public enum JamsilStadiumStatusType {
+public enum JamsilStadiumStatusType implements StadiumStatusType{
 
     RED("레드석",
             List.of("응원도 적당히 즐길 수 있지만, 야구나 함께 온 동행자와의 대화에도 집중할 수 있는 구역이에요!"),
@@ -202,69 +202,4 @@ public enum JamsilStadiumStatusType {
     private final List<String> page3Keywords;
     private final List<String> forbiddenKeywords;
 
-    public static int getMatchingKeywordCount(List<String> keywords, List<String> clientKeywords){
-        return (int) keywords.stream()
-                .filter(clientKeywords::contains)
-                .count();
-    }
-
-    public static boolean hasForbiddenKeywords(List<String> forbiddenKeywords, List<String> clientKeywords){
-        return forbiddenKeywords.stream()
-                .anyMatch(clientKeywords::contains);
-    }
-
-    public static List<Map<String, Object>> getTopRankedZones(List<String> clientKeywords){
-        return Arrays.stream(JamsilStadiumStatusType.values())
-                .map(zone -> {
-                    if (hasForbiddenKeywords(zone.getForbiddenKeywords(), clientKeywords)){
-                        return null;
-                    }
-
-                    int page1Count = getMatchingKeywordCount(zone.getPage1Keywords(), clientKeywords);
-                    int page2Count = getMatchingKeywordCount(zone.getPage2Keywords(), clientKeywords);
-                    int page3Count = getMatchingKeywordCount(zone.getPage3Keywords(), clientKeywords);
-
-                    // 총 겹치는 개수 계산
-                    int totalMatchCount = page1Count + page2Count + page3Count;
-
-                    Map<String, Object> result = new HashMap<>();
-                    result.put("stadium", "잠실");
-                    result.put("zone", zone.zone);
-                    result.put("explanations", zone.getExplanations());
-                    result.put("tip", zone.tip);
-                    result.put("references", zone.getReferences());
-                    result.put("totalMatchCount", totalMatchCount);
-                    result.put("page1Count", page1Count);
-                    result.put("page2Count", page2Count);
-                    result.put("page3Count", page3Count);
-                    return result;
-                })
-                .filter(Objects::nonNull)
-                .filter(result -> (int) result.get("totalMatchCount") > 0)
-                .sorted((a, b) -> {
-                    int totalCompare = Integer.compare((int) b.get("totalMatchCount"), (int) a.get("totalMatchCount"));
-                    if (totalCompare == 0) {
-                        int page2Compare = Integer.compare((int) b.get("page2Count"), (int) a.get("page2Count"));
-                        if (page2Compare == 0) {
-                            int page3Compare = Integer.compare((int) b.get("page3Count"), (int) a.get("page3Count"));
-                            if (page3Compare == 0) {
-                                return Integer.compare((int) b.get("page1Count"), (int) a.get("page1Count"));
-                            }
-                            return page3Compare;
-                        }
-                        return page2Compare;
-                    }
-                    return totalCompare;
-                })
-                .limit(3)
-                .map(result -> {
-                    log.info("zone: {}, totalMatchCount: {}, page1Count: {}, page2Count: {}, page3Count: {}", result.get("zone"), result.get("totalMatchCount"), result.get("page1Count"), result.get("page2Count"), result.get("page3Count"));
-                    result.remove("totalMatchCount");
-                    result.remove("page1Count");
-                    result.remove("page2Count");
-                    result.remove("page3Count");
-                    return result;
-                })
-                .toList();
-    }
 }
