+++ src/main/java/kusitms/backend/stadium/presentation/StadiumController.java
@@ -1,5 +1,6 @@
 package kusitms.backend.stadium.presentation;
 
+import kusitms.backend.stadium.common.TopRankedZones;
 import kusitms.backend.stadium.domain.entity.enums.JamsilStadiumStatusType;
 import kusitms.backend.stadium.dto.request.TopRankedZoneRequestDto;
 import lombok.RequiredArgsConstructor;
@@ -9,6 +10,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequiredArgsConstructor
@@ -17,6 +19,9 @@ public class StadiumController {
 
     @PostMapping("/zone")
     public Object serveZones(@RequestBody TopRankedZoneRequestDto request){
-        return JamsilStadiumStatusType.getTopRankedZones(List.of(request.clientKeywords()));
+        List<Map<String, Object>> jamsilTopZones = TopRankedZones.getTopRankedZones(
+                JamsilStadiumStatusType.values(), "잠실", List.of(request.clientKeywords()));
+
+        return jamsilTopZones;
     }
 }
